test: split oversized adoption test module and fix legacy auto_merge seed
test_pr_monitor_adoption_part_001.py was 1509 lines, over the 1500-line
first-party maintainability guard. Move the terminal-adoption / lineage
test methods into a new sibling test_pr_monitor_adoption_part_004.py
(class TestPullRequestMonitorAdoptionServicePart004), preserving behavior.

Also fix test_replay_treats_legacy_non_object_task_policy_as_empty_agent_policy:
the adoption auto_merge matcher reconstructs a legacy row's intent from the
persisted column, so when simulating a pre-change legacy row (wiped intent
key) set the column to True to model the historical omitted-request default,
matching the omitted replay instead of spuriously raising a policy conflict.
